Move /jest into /config
