+++ jest.config.ts
@@ -8,7 +8,7 @@ const config: Config = {
         '!**/vendor/**',
         '!**/test-connection.ts',
         '!jest.config.ts',
-        '!**/jest/**',
+        '!**/config/**',
         '!eslint.config.ts',
         '!**/data/**',
         '!src/database/connection/mongoose.ts',
@@ -22,11 +22,11 @@ const config: Config = {
             statements: 90,
         },
     },
-    globalSetup: '<rootDir>/jest/globalSetup.ts',
-    globalTeardown: '<rootDir>/jest/globalTeardown.ts',
+    globalSetup: '<rootDir>/config/jest/globalSetup.ts',
+    globalTeardown: '<rootDir>/config/jest/globalTeardown.ts',
     maxWorkers: 1,
     preset: 'ts-jest',
-    setupFilesAfterEnv: ['<rootDir>/jest/setupFile.ts'],
+    setupFilesAfterEnv: ['<rootDir>/config/jest/setupFile.ts'],
     silent: true,
     testMatch: ['<rootDir>/src/**/*.test.ts'],
     testTimeout: 30000,
